+++ bundle.js
@@ -4,6 +4,8 @@ import "./js/inputmask.date.extensions";
 import "./js/inputmask.extensions";
 import "./js/inputmask.numeric.extensions";
 import "./js/inputmask.phone.extensions";
+//remove me later
+import "./js/inputmask.regex.extensions";
 
 // import "./js/phone-codes/phone-be";
 // import "./js/phone-codes/phone-nl";
@@ -15,7 +17,7 @@ import im from "./js/inputmask.js";
 import dependencyLib from "./js/dependencyLibs/inputmask.dependencyLib";
 import jQuery from "jquery";
 if (dependencyLib === jQuery) {
-    require("./js/jquery.inputmask");
+    require("./js/integrations/jquery.inputmask");
 }
 window.Inputmask = im;