+++ js/integrations/jquery.inputmask.js
@@ -8,9 +8,9 @@
 
 (function (factory) {
 	if (typeof define === "function" && define.amd) {
-		define(["jquery", "./inputmask"], factory);
+		define(["jquery", "../inputmask"], factory);
 	} else if (typeof exports === "object") {
-		module.exports = factory(require("jquery"), require("./inputmask"));
+		module.exports = factory(require("jquery"), require("../inputmask"));
 	} else {
 		factory(jQuery, window.Inputmask);
 	}