+++ qunit/index.js
@@ -11,7 +11,7 @@ import "../js/phone-codes/phone-nl";
 import "../js/phone-codes/phone-ru";
 import "../js/phone-codes/phone-uk";
 import "../js/phone-codes/phone";
-import "../js/jquery.inputmask";
+import "../js/integrations/jquery.inputmask";
 
 import  Inputmask from "../js/inputmask";
 import dependencyLib from "../js/dependencyLibs/inputmask.dependencyLib";
@@ -25,10 +25,10 @@ import simulator from "./simulator.js";
 
 
 // android testing
-Inputmask.extendDefaults({
-   inputEventOnly: true
-});
-window.Inputmask = Inputmask; //inject globally for the simulator to detect inputeventonly
+// Inputmask.extendDefaults({
+//    inputEventOnly: true
+// });
+// window.Inputmask = Inputmask; //inject globally for the simulator to detect inputeventonly
 
 import tests_alternations from "./tests_alternations";
 import tests_attributes from "./tests_attributes";